Moved shop data into firestore database
Moved the shop data from a file on the front end to our firebase
database to make the data more available and easier to maintain.

Created and implemented the WithSpinner HOC for the Shop Page, to
allow time for the data to be retrieved and inserted into its
Redux reducer so that the page can be populated properly.